baremetal: Fix detection for Keil toolchains
Previous implementation does not work if qbs is compiled with
64-bit compiler. The reason is the wrong path in the registry,
because it differs for 32/64 bit applications.

Besides, the search in the registry was difficult, it was done
through the `Uninstall` section, with the parsing of the
`tools.ini` file.

But we can use an alternative search algorithm using the
`HKLM/SOFTWARE` path instead, that simplifies the algorithm.

This patch tested with the Qbs compiled with both 32/64 bit
compilers.

Change-Id: I87e9f62966482ba840b0aa92d4600be29b296384